data-pipeline/requirements.txt
# AI RAG Pipeline Requirements
# Core dependencies for AI article scraping and vectorization

# ==================== Core Framework ====================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1

# ==================== Web Scraping ====================
playwright==1.40.0
beautifulsoup4==4.12.0
lxml==4.9.3

# ==================== PDF Processing ====================
PyMuPDF==1.23.8

# ==================== Vector Database and Embeddings ====================
pinecone-client==3.0.0
google-generativeai==0.3.1
tiktoken==0.5.1

# ==================== AWS & Cloud Storage ====================
boto3==1.34.0

# ==================== Utility Libraries ====================
tqdm==4.66.1
colorama==0.4.6
python-dateutil==2.8.2

# ==================== Optional Dependencies ====================
# Uncomment below if you need testing features:
# pytest==7.4.3
# pytest-asyncio==0.21.1
